+++ b4n-tui/table/header.rs
@@ -12,8 +12,6 @@ pub struct Header {
     name: Column,                         // column: 1
     age: Column,                          // column: extra_columns len + 2 (last column)
     extra_columns: Option<Box<[Column]>>, // columns: 2 .. n
-    extra_columns_text: Option<String>,
-    double_spaces_count: usize,
     all_extra_width: usize,
     extra_space: usize,
     sort_symbols: Rc<[char]>,
@@ -40,8 +38,6 @@ impl Header {
             name: NAME,
             age: AGE,
             extra_columns,
-            extra_columns_text: None,
-            double_spaces_count: 0,
             all_extra_width: extra_width,
             extra_space,
             sort_symbols,
@@ -109,13 +105,12 @@ impl Header {
 
     /// Returns the number of doubled spaces for additional columns.
     pub fn double_spaces_count(&self) -> usize {
-        self.double_spaces_count
+        self.cache.double_spaces_count
     }
 
     /// Recalculates extra columns text and width.
     pub fn recalculate_extra_columns(&mut self) {
         self.cache.invalidate();
-        self.extra_columns_text = None;
         self.all_extra_width = get_extra_columns_len(self.extra_columns.as_deref()) + 9; // AGE + all spaces = 9
         self.extra_space = get_extra_space(self.extra_columns.as_deref());
     }
@@ -123,7 +118,6 @@ impl Header {
     /// Resets `data_len` in each not fixed column.
     pub fn reset_data_lengths(&mut self) {
         self.cache.invalidate();
-        self.extra_columns_text = None;
         self.group.data_len = 0;
         self.name.data_len = 0;
         if let Some(columns) = &mut self.extra_columns {
@@ -143,7 +137,6 @@ impl Header {
     /// Sets data length for the provided column.
     pub fn set_data_length(&mut self, column_no: usize, new_data_len: usize) {
         self.cache.invalidate();
-        self.extra_columns_text = None;
         if let Some(column) = self.column_mut(column_no)
             && !column.is_fixed
         {
@@ -160,8 +153,10 @@ impl Header {
     pub fn refresh_text(&mut self, view: ViewType, width: usize) {
         if self.cache.area_width.is_none_or(|w| w != width) || self.cache.view != view {
             let (group_width, name_width, _) = self.get_widths(view, width);
-            self.update_extra_columns_text(name_width);
-            self.cache.text = self.get_text_string(view, group_width, name_width.saturating_sub(self.double_spaces_count), width);
+            self.update_cached_extra_columns_text(name_width);
+
+            let spaces_width = self.cache.double_spaces_count;
+            self.cache.text = self.get_text_string(view, group_width, name_width.saturating_sub(spaces_width), width);
             self.cache.view = view;
             self.cache.area_width = Some(width);
             self.cache.text_length = Some(self.cache.text.chars().count());
@@ -278,7 +273,7 @@ impl Header {
         header.push(' ');
         header.push_column(&self.name, name_width, self.is_sorted_descending);
         header.push(' ');
-        header.push_str(self.extra_columns_text.as_deref().unwrap_or_default());
+        header.push_str(self.cache.extra_columns_text());
         header.push(' ');
         header.push_column(&self.age, self.age.max_len(), self.is_sorted_descending);
         header.push(' ');
@@ -286,7 +281,7 @@ impl Header {
         header
     }
 
-    fn update_extra_columns_text(&mut self, name_width: usize) {
+    fn update_cached_extra_columns_text(&mut self, name_width: usize) {
         let double_spaces_count = self
             .extra_columns
             .as_ref()
@@ -300,12 +295,12 @@ impl Header {
             false
         };
 
-        self.extra_columns_text = Some(get_extra_columns_text(
+        self.cache.extra_columns_text = Some(get_extra_columns_text(
             self.extra_columns.as_deref(),
             self.is_sorted_descending,
             has_enough_space,
         ));
-        self.double_spaces_count = if has_enough_space { double_spaces_count } else { 0 };
+        self.cache.double_spaces_count = if has_enough_space { double_spaces_count } else { 0 };
     }
 
     fn column(&self, column_no: usize) -> Option<&Column> {
@@ -362,6 +357,8 @@ struct HeaderCache {
     pub view: ViewType,
     pub area_width: Option<usize>,
     pub text_length: Option<usize>,
+    pub extra_columns_text: Option<String>,
+    pub double_spaces_count: usize,
 }
 
 impl HeaderCache {
@@ -369,6 +366,13 @@ impl HeaderCache {
     pub fn invalidate(&mut self) {
         self.area_width = None;
         self.text_length = None;
+        self.extra_columns_text = None;
+        self.double_spaces_count = 0;
+    }
+
+    /// Returns the cached extra columns text.
+    pub fn extra_columns_text(&self) -> &str {
+        self.extra_columns_text.as_deref().unwrap_or_default()
     }
 }